"Updated the Login XML file to add ID's for each of the clickable items and changed its name to login_screen
Updated the java name to Login_Screen
Wrote the code to check for authentication email and password, and listeners to check if any of the pushable buttons have been pressed so you can move to the next screen. Authentication will be checked using Google Firebase.

Login Page should Be done. 5-22-23